move use-token-toggle-details to utils folder
