+++ src/utils/use-token-toggle-details.ts
